Move all ACL samples into separate files
