+++ google-cloud-storage/samples/acls.rb
@@ -12,201 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-def print_bucket_acl bucket_name:
-  # [START storage_print_bucket_acl]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  puts "ACL for #{bucket_name}:"
-
-  bucket.acl.owners.each do |owner|
-    puts "OWNER #{owner}"
-  end
-
-  bucket.acl.writers.each do |writer|
-    puts "WRITER #{writer}"
-  end
-
-  bucket.acl.readers.each do |reader|
-    puts "READER #{reader}"
-  end
-  # [END storage_print_bucket_acl]
-end
-
-def print_bucket_acl_for_user bucket_name:, email:
-  # [START storage_print_bucket_acl_for_user]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  puts "Permissions for #{email}:"
-  puts "OWNER"  if bucket.acl.owners.include?  email
-  puts "WRITER" if bucket.acl.writers.include? email
-  puts "READER" if bucket.acl.readers.include? email
-  # [END storage_print_bucket_acl_for_user]
-end
-
-def add_bucket_owner bucket_name:, email:
-  # [START storage_add_bucket_owner]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  bucket.acl.add_owner email
-
-  puts "Added OWNER permission for #{email} to #{bucket_name}"
-  # [END storage_add_bucket_owner]
-end
-
-def remove_bucket_acl bucket_name:, email:
-  # [START storage_remove_bucket_owner]
-  # project_id  = "Your Google Cloud project ID"
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  bucket.acl.delete email
-
-  puts "Removed ACL permissions for #{email} from #{bucket_name}"
-  # [END storage_remove_bucket_owner]
-end
-
-def add_bucket_default_owner bucket_name:, email:
-  # [START storage_add_bucket_default_owner]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  bucket.default_acl.add_owner email
-
-  puts "Added default OWNER permission for #{email} to #{bucket_name}"
-  # [END storage_add_bucket_default_owner]
-end
-
-def remove_bucket_default_acl bucket_name:, email:
-  # [START storage_remove_bucket_default_owner]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-
-  bucket.default_acl.delete email
-
-  puts "Removed default ACL permissions for #{email} from #{bucket_name}"
-  # [END storage_remove_bucket_default_owner]
-end
-
-def print_file_acl bucket_name:, file_name:
-  # [START storage_print_file_acl]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # file_name   = "Name of a file in the Storage bucket"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-  file    = bucket.file file_name
-
-  puts "ACL for #{file_name} in #{bucket_name}:"
-
-  file.acl.owners.each do |owner|
-    puts "OWNER #{owner}"
-  end
-
-  file.acl.readers.each do |reader|
-    puts "READER #{reader}"
-  end
-  # [END storage_print_file_acl]
-end
-
-def print_file_acl_for_user bucket_name:, file_name:, email:
-  # [START storage_print_file_acl_for_user]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # file_name   = "Name of a file in the Storage bucket"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-  file    = bucket.file file_name
-
-  puts "Permissions for #{email}:"
-  puts "OWNER"  if file.acl.owners.include?  email
-  puts "READER" if file.acl.readers.include? email
-  # [END storage_print_file_acl_for_user]
-end
-
-def add_file_owner bucket_name:, file_name:, email:
-  # [START storage_add_file_owner]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # file_name   = "Name of a file in the Storage bucket"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-  file    = bucket.file file_name
-
-  file.acl.add_owner email
-
-  puts "Added OWNER permission for #{email} to #{file_name}"
-  # [END storage_add_file_owner]
-end
-
-def remove_file_acl bucket_name:, file_name:, email:
-  # [START storage_remove_file_owner]
-  # The ID of your GCS bucket
-  # bucket_name = "your-unique-bucket-name"
-  # file_name   = "Name of a file in the Storage bucket"
-  # email       = "Google Cloud Storage ACL Entity email"
-
-  require "google/cloud/storage"
-
-  storage = Google::Cloud::Storage.new
-  bucket  = storage.bucket bucket_name
-  file    = bucket.file file_name
-
-  file.acl.delete email
-
-  puts "Removed ACL permissions for #{email} from #{file_name}"
-  # [END storage_remove_file_owner]
-end
-
 def run_sample arguments
   command = arguments.shift
